Moving SPE files
