+++ client/blocks/upe/spe/handle-display-of-saving-checkbox.js
@@ -1,10 +1,10 @@
 import { NON_RECURRING_METHODS } from 'wcstripe/stripe-utils/constants';
 
-export const handleDisplayOfSavingCheckboxForSpe = ( method ) => {
+export const handleDisplayOfSavingCheckbox = ( method ) => {
 	const saveCardInfoContainer = document.querySelector(
 		'.wc-block-components-payment-methods__save-card-info'
 	);
-	if ( saveCardInfoContainer.length > 0 ) {
+	if ( saveCardInfoContainer ) {
 		saveCardInfoContainer.style.display = NON_RECURRING_METHODS.includes(
 			method
 		)
