+++ client/blocks/upe/upe-deferred-intent-creation/payment-processor.js
@@ -22,8 +22,8 @@ import {
 } from 'wcstripe/stripe-utils/cash-app-limit-notice-handler';
 import { isLinkEnabled } from 'wcstripe/stripe-utils';
 import { PAYMENT_METHOD_CASHAPP } from 'wcstripe/stripe-utils/constants';
-import { applySinglePaymentElementStyles } from 'wcstripe/blocks/upe/apply-single-payment-element-styles';
-import { handleDisplayOfSavingCheckboxForSpe } from 'wcstripe/blocks/upe/handle-display-of-saving-checkbox-for-spe';
+import { applySinglePaymentElementStyles } from 'wcstripe/blocks/upe/spe/apply-single-payment-element-styles';
+import { handleDisplayOfSavingCheckbox } from 'wcstripe/blocks/upe/spe/handle-display-of-saving-checkbox';
 
 const noop = () => null;
 
@@ -326,7 +326,7 @@ const PaymentProcessor = ( {
 		setSelectedPaymentMethodType( value.type );
 		setIsPaymentElementComplete( complete );
 		if ( getBlocksConfiguration()?.isSPEEnabled ) {
-			handleDisplayOfSavingCheckboxForSpe( value.type );
+			handleDisplayOfSavingCheckbox( value.type );
 		}
 	};
